CETERRA 65 uint purge

+++ Assets/Code/Scripts/Domains/Board/Field/CatastropheFieldBehaviour.cs
@@ -37,7 +37,7 @@ public class CatastropheFieldBehaviour : FieldBehaviour {
                 int positionYIterationPlayer = currentPlayer.SplineKnotIndex.Knot;
                 if (positionXCurrentPlayer - effectRadius <= positionXIterationPlayer && positionXCurrentPlayer + effectRadius >= positionXIterationPlayer) {
                     if (positionYCurrentPlayer - effectRadius <= positionYIterationPlayer && positionYCurrentPlayer + effectRadius >= positionYIterationPlayer) {
-                        currentPlayer.RemoveHealth((uint)healthEffect);
+                        currentPlayer.RemoveHealth(healthEffect);
                         Debug.Log($"Player {currentPlayer} got health removed by amount {healthEffect}");
                     }
                 }
